pyproject.toml
[project]
name = "gpr-viewer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = ["h5py>=3.13.0", "numpy>=2.2.2", "pillow>=11.1.0"]
